[State Sync] Remove SynchronizerEnvHelper and merge functions into SynchronizerEnv
Closes: #7198
